# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seoul 119 Fire Risk Forecast API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RiskLevel = typeof RiskLevel[keyof typeof RiskLevel];


export const RiskLevel = {
  관심: '관심',
  주의: '주의',
  경계: '경계',
  위험: '위험',
} as const;

/**
 * KMA=기상청 연동, NO_KEY=키 미등록, ERROR=연결 실패, DB_FALLBACK=샘플 데이터
 */
export type WeatherSource = typeof WeatherSource[keyof typeof WeatherSource];


export const WeatherSource = {
  KMA: 'KMA',
  NO_KEY: 'NO_KEY',
  ERROR: 'ERROR',
  DB_FALLBACK: 'DB_FALLBACK',
} as const;

export interface TodayRiskSummary {
  /** Overall Seoul fire risk score 0-100 */
  overallScore: number;
  riskLevel: RiskLevel;
  mainFireType: string;
  topRiskFactors: string[];
  preventionMessage: string;
  date: string;
  highRiskDistrictCount: number;
}

export interface DistrictRisk {
  name: string;
  score: number;
  riskLevel: RiskLevel;
  mainFireType: string;
  dangerTime: string;
  weatherFactor?: string;
  preventionMessage?: string;
  lat: number;
  lng: number;
}

export interface ScoreBreakdown {
  historicalFrequency: number;
  regionalHistory: number;
  causePersistence: number;
  weatherRisk: number;
  regionRisk: number;
}

export interface DistrictRiskDetail {
  name: string;
  score: number;
  riskLevel: RiskLevel;
  mainFireType: string;
  dangerTime: string;
  weatherFactor: string;
  regionFactors: string[];
  citizenAdvice: string[];
  firefighterAdvice: string[];
  lat: number;
  lng: number;
  scoreBreakdown?: ScoreBreakdown;
}

export interface WeatherCurrentResponse {
  temperature?: number | null;
  humidity?: number | null;
  windSpeed?: number | null;
  windDirection?: string | null;
  precipitation?: number | null;
  skyCondition?: string | null;
  precipitationType?: string | null;
  source: WeatherSource;
  sourceMessage: string;
  observedAt?: string | null;
}

export interface DayForecast {
  date: string;
  temperature: number;
  maxTemp: number;
  minTemp: number;
  humidity: number;
  precipitationProb: number;
  precipitation?: number;
  windSpeed: number;
  windDirection?: string;
  skyCondition: string;
  precipitationType?: string;
  isDry: boolean;
  isHeatwave: boolean;
  isColdwave: boolean;
  isStrongWind: boolean;
  riskFactors?: string[];
  /** 기상 위험 점수 (0-20) */
  weatherScore?: number;
  source: WeatherSource;
  sourceMessage: string;
}

export interface WeatherForecast {
  today: DayForecast;
  tomorrow: DayForecast;
  source: WeatherSource;
  sourceMessage: string;
}

export type AlertMessageTargetAudience = typeof AlertMessageTargetAudience[keyof typeof AlertMessageTargetAudience];


export const AlertMessageTargetAudience = {
  citizen: 'citizen',
  firefighter: 'firefighter',
  all: 'all',
} as const;

export interface AlertMessage {
  id: number;
  title: string;
  message: string;
  severity: RiskLevel;
  category: string;
  targetAudience: AlertMessageTargetAudience;
  timeRange?: string;
  district?: string | null;
}

export interface MonthlyStatistic {
  month: number;
  monthName?: string;
  count: number;
  avgScore: number;
  season?: string;
}

export interface HourlyStatistic {
  hour: number;
  label?: string;
  count: number;
  percentage: number;
}

export interface FireCauseStat {
  cause: string;
  count: number;
  percentage: number;
}

export interface SeasonalStatistic {
  season: string;
  count: number;
  mainCause: string;
  riskLevel: RiskLevel;
}

export interface MapDistrict {
  name: string;
  lat: number;
  lng: number;
  score: number;
  riskLevel: RiskLevel;
  mainFireType?: string;
  hasTraditionalMarket?: boolean;
  hasForestArea?: boolean;
  hasOldBuildings?: boolean;
}

export interface RiskTypeCount {
  type: string;
  count: number;
}

export interface PatrolRecommendation {
  district: string;
  priority: number;
  reason: string;
  action: string;
  timeRange?: string;
}

export interface FirefighterDashboard {
  date: string;
  totalHighRiskAreas: number;
  priorityDistricts: DistrictRisk[];
  riskTypeBreakdown: RiskTypeCount[];
  patrolRecommendations: PatrolRecommendation[];
}

export type GetWeatherCurrentParams = {
/**
 * 자치구명 (예: 성동구). 없으면 서울 기본 격자 사용.
 */
district?: string;
};

export type GetWeatherForecastParams = {
/**
 * 자치구명 (예: 성동구)
 */
district?: string;
/**
 * 기상청 격자 X 좌표 (기본값 60)
 */
nx?: number;
/**
 * 기상청 격자 Y 좌표 (기본값 127)
 */
ny?: number;
};

export type GetTodayAlertsParams = {
userType?: GetTodayAlertsUserType;
};

export type GetTodayAlertsUserType = typeof GetTodayAlertsUserType[keyof typeof GetTodayAlertsUserType];


export const GetTodayAlertsUserType = {
  citizen: 'citizen',
  firefighter: 'firefighter',
} as const;