# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Seoul 119 Fire Risk Forecast API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /risk/districts/{districtName}:
    get:
      operationId: getRiskByDistrict
      tags: [risk]
      summary: Get fire risk for a specific district
      parameters:
        - name: districtName
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: District risk detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DistrictRiskDetail"
        "404":
          description: District not found
...
  /weather/current:
    get:
      operationId: getWeatherCurrent
      tags: [weather]
      summary: Get current real-time weather (초단기실황)
      parameters:
        - name: district
          in: query
          required: false
          schema:
            type: string
          description: "자치구명 (예: 성동구). 없으면 서울 기본 격자 사용."
      responses:
        "200":
          description: Current weather conditions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherCurrentResponse"
  /weather/forecast:
    get:
      operationId: getWeatherForecast
      tags: [weather]
      summary: Get today and tomorrow weather forecast (단기예보)
      parameters:
        - name: district
          in: query
          required: false
          schema:
            type: string
          description: "자치구명 (예: 성동구)"
        - name: nx
          in: query
          required: false
          schema:
            type: integer
          description: 기상청 격자 X 좌표 (기본값 60)
        - name: ny
          in: query
          required: false
          schema:
            type: integer
          description: 기상청 격자 Y 좌표 (기본값 127)
      responses:
        "200":
          description: Weather forecast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherForecast"
...
components:
  schemas:
...
    RiskLevel:
      type: string
      enum: [관심, 주의, 경계, 위험]
    WeatherSource:
      type: string
      enum: [KMA, NO_KEY, ERROR, DB_FALLBACK]
      description: "KMA=기상청 연동, NO_KEY=키 미등록, ERROR=연결 실패, DB_FALLBACK=샘플 데이터"
...
    DistrictRiskDetail:
      type: object
      required: [name, score, riskLevel, mainFireType, dangerTime, weatherFactor, regionFactors, citizenAdvice, firefighterAdvice, lat, lng]
      properties:
        name:
          type: string
        score:
          type: integer
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        mainFireType:
          type: string
        dangerTime:
          type: string
        weatherFactor:
          type: string
        regionFactors:
          type: array
          items:
            type: string
        citizenAdvice:
          type: array
          items:
            type: string
        firefighterAdvice:
          type: array
          items:
            type: string
        lat:
          type: number
        lng:
          type: number
        scoreBreakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
    ScoreBreakdown:
      type: object
      required: [historicalFrequency, regionalHistory, causePersistence, weatherRisk, regionRisk]
      properties:
        historicalFrequency:
          type: integer
        regionalHistory:
          type: integer
        causePersistence:
          type: integer
        weatherRisk:
          type: integer
        regionRisk:
          type: integer
    WeatherCurrentResponse:
      type: object
      required:
        - source
        - sourceMessage
      properties:
        temperature:
          type: number
          nullable: true
        humidity:
          type: integer
          nullable: true
        windSpeed:
          type: number
          nullable: true
        windDirection:
          type: string
          nullable: true
        precipitation:
          type: number
          nullable: true
        skyCondition:
          type: string
          nullable: true
        precipitationType:
          type: string
          nullable: true
        source:
          $ref: "#/components/schemas/WeatherSource"
        sourceMessage:
          type: string
        observedAt:
          type: string
          nullable: true
    WeatherForecast:
      type: object
      required: [today, tomorrow, source, sourceMessage]
      properties:
        today:
          $ref: "#/components/schemas/DayForecast"
        tomorrow:
          $ref: "#/components/schemas/DayForecast"
        source:
          $ref: "#/components/schemas/WeatherSource"
        sourceMessage:
          type: string
    DayForecast:
      type: object
      required:
        - date
        - temperature
        - maxTemp
        - minTemp
        - humidity
        - precipitationProb
        - windSpeed
        - skyCondition
        - isDry
        - isHeatwave
        - isColdwave
        - isStrongWind
        - source
        - sourceMessage
      properties:
        date:
          type: string
        temperature:
          type: number
        maxTemp:
          type: number
        minTemp:
          type: number
        humidity:
          type: integer
        precipitationProb:
          type: integer
        precipitation:
          type: number
        windSpeed:
          type: number
        windDirection:
          type: string
        skyCondition:
          type: string
        precipitationType:
          type: string
        isDry:
          type: boolean
        isHeatwave:
          type: boolean
        isColdwave:
          type: boolean
        isStrongWind:
          type: boolean
        riskFactors:
          type: array
          items:
            type: string
        weatherScore:
          type: integer
          description: 기상 위험 점수 (0-20)
        source:
          $ref: "#/components/schemas/WeatherSource"
        sourceMessage:
          type: string
...