# lib/api-zod/src/generated/types/dayForecast.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seoul 119 Fire Risk Forecast API
 * OpenAPI spec version: 0.1.0
 */
import type { WeatherSource } from './weatherSource';

export interface DayForecast {
  date: string;
  temperature: number;
  maxTemp: number;
  minTemp: number;
  humidity: number;
  precipitationProb: number;
  precipitation?: number;
  windSpeed: number;
  windDirection?: string;
  skyCondition: string;
  precipitationType?: string;
  isDry: boolean;
  isHeatwave: boolean;
  isColdwave: boolean;
  isStrongWind: boolean;
  riskFactors?: string[];
  /** 기상 위험 점수 (0-20) */
  weatherScore?: number;
  source: WeatherSource;
  sourceMessage: string;
}