# lib/api-zod/src/generated/types/weatherCurrentResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seoul 119 Fire Risk Forecast API
 * OpenAPI spec version: 0.1.0
 */
import type { WeatherSource } from './weatherSource';

export interface WeatherCurrentResponse {
  temperature?: number | null;
  humidity?: number | null;
  windSpeed?: number | null;
  windDirection?: string | null;
  precipitation?: number | null;
  skyCondition?: string | null;
  precipitationType?: string | null;
  source: WeatherSource;
  sourceMessage: string;
  observedAt?: string | null;
}