# lib/api-zod/src/generated/types/weatherForecast.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seoul 119 Fire Risk Forecast API
 * OpenAPI spec version: 0.1.0
 */
import type { DayForecast } from './dayForecast';
import type { WeatherSource } from './weatherSource';

export interface WeatherForecast {
  today: DayForecast;
  tomorrow: DayForecast;
  source: WeatherSource;
  sourceMessage: string;
}