# lib/api-zod/src/generated/types/weatherSource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Seoul 119 Fire Risk Forecast API
 * OpenAPI spec version: 0.1.0
 */

/**
 * KMA=기상청 연동, NO_KEY=키 미등록, ERROR=연결 실패, DB_FALLBACK=샘플 데이터
 */
export type WeatherSource = typeof WeatherSource[keyof typeof WeatherSource];


export const WeatherSource = {
  KMA: 'KMA',
  NO_KEY: 'NO_KEY',
  ERROR: 'ERROR',
  DB_FALLBACK: 'DB_FALLBACK',
} as const;